moved components into folders (first draft)
