+++ src/components/modals/AddFossilModal.tsx
@@ -1,48 +1,57 @@
 import React, { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { STONES } from "@/src/services/init";
+import { FOSSILS } from "@/src/services/init.ts";
 import { FiX, FiInfo } from "react-icons/fi";
-import LocationSuggestionInput from "./LocationSuggestionInput";
-import Tooltip from "./Tooltip";
-import { useFocusTrap } from "@/src/hooks/useFocusTrap";
+import LocationSuggestionInput from "../LocationSuggestionInput.tsx";
+import Tooltip from "@/src/components/basic/Tooltip.tsx";
+import { useFocusTrap } from "@/src/hooks/useFocusTrap.ts";
 import {
   focusRingCardClasses,
   focusRingClasses,
 } from "@/src/styles/focusRing.ts";
 
-interface AddStoneModalProps {
+interface AddFossilModalProps {
   isOpen: boolean;
   onClose: () => void;
-  onAdd: (stoneId: string, location: string, inBag: boolean) => void;
+  onAdd: (fossilId: string, location: string, inBag: boolean) => void;
   maxGeneration: number;
   alreadyOwnedIds: string[];
+  infiniteFossilsEnabled: boolean;
   gameVersionId?: string;
 }
 
-const AddStoneModal: React.FC<AddStoneModalProps> = ({
+const AddFossilModal: React.FC<AddFossilModalProps> = ({
   isOpen,
   onClose,
   onAdd,
   maxGeneration,
+  alreadyOwnedIds,
+  infiniteFossilsEnabled,
   gameVersionId,
 }) => {
   const { t } = useTranslation();
   const { containerRef } = useFocusTrap(isOpen);
-  const [selectedStoneId, setSelectedStoneId] = useState("");
+  const [selectedFossilId, setSelectedFossilId] = useState("");
   const [location, setLocation] = useState("");
   const [inBag, setInBag] = useState(true);
 
-  const availableStones = useMemo(() => {
-    return STONES.filter((s) => s.gen <= maxGeneration);
-  }, [maxGeneration]);
+  const availableFossils = useMemo(() => {
+    return FOSSILS.filter((f) => {
+      const genMatch = f.gen <= maxGeneration;
+      if (!infiniteFossilsEnabled) {
+        return genMatch && !alreadyOwnedIds.includes(f.id);
+      }
+      return genMatch;
+    });
+  }, [maxGeneration, alreadyOwnedIds, infiniteFossilsEnabled]);
 
   if (!isOpen) return null;
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    if (!selectedStoneId) return;
-    onAdd(selectedStoneId, inBag ? "" : location, inBag);
-    setSelectedStoneId("");
+    if (!selectedFossilId) return;
+    onAdd(selectedFossilId, inBag ? "" : location, inBag);
+    setSelectedFossilId("");
     setLocation("");
     setInBag(true);
   };
@@ -58,7 +67,7 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
       >
         <div className="flex justify-between items-center mb-4">
           <h2 className="text-xl font-bold dark:text-gray-100">
-            {t("modals.addStone.title")}
+            {t("modals.addFossil.title")}
           </h2>
           <button
             onClick={onClose}
@@ -71,37 +80,37 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
             <label className="block text-sm font-bold text-gray-700 dark:text-gray-300 mb-2">
-              {t("modals.addStone.stoneLabel")}
+              {t("modals.addFossil.fossilLabel")}
             </label>
 
-            {availableStones.length > 0 ? (
+            {availableFossils.length > 0 ? (
               <div className="grid grid-cols-3 gap-2 p-1">
-                {availableStones.map((s) => (
+                {availableFossils.map((f) => (
                   <button
-                    key={s.id}
+                    key={f.id}
                     type="button"
-                    onClick={() => setSelectedStoneId(s.id)}
+                    onClick={() => setSelectedFossilId(f.id)}
                     className={`flex flex-col items-center p-2 rounded-md border transition-all ${focusRingCardClasses} ${
-                      selectedStoneId === s.id
+                      selectedFossilId === f.id
                         ? "border-green-500 bg-green-50 dark:bg-green-900/20"
                         : "border-gray-200 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-700"
                     }`}
                   >
                     <img
-                      src={`/stone-sprites/${s.sprite}`}
-                      alt={s.id}
+                      src={`/fossil-sprites/${f.sprite}`}
+                      alt={f.id}
                       className="w-10 h-10 object-contain"
                       style={{ imageRendering: "pixelated" }}
                     />
                     <span className="text-[10px] mt-1 text-center dark:text-gray-200">
-                      {t(`stones.${s.id}`)}
+                      {t(`fossils.${f.id}`)}
                     </span>
                   </button>
                 ))}
               </div>
             ) : (
               <div className="p-4 rounded-md border border-amber-200 dark:border-amber-800 bg-amber-50 dark:bg-amber-900/20 text-xs text-amber-800 dark:text-amber-200 leading-relaxed">
-                {t("modals.addStone.emptyListExplanation")}
+                {t("modals.addFossil.emptyListExplanation")}
               </div>
             )}
           </div>
@@ -109,40 +118,40 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
           <div className="flex items-center gap-2 py-2">
             <input
               type="checkbox"
-              id="stoneInBag"
+              id="inBag"
               checked={inBag}
               onChange={(e) => {
                 setInBag(e.target.checked);
                 if (e.target.checked) setLocation("");
               }}
-              disabled={availableStones.length === 0}
+              disabled={availableFossils.length === 0}
               className="h-4 w-4 accent-green-600 cursor-pointer disabled:cursor-not-allowed"
             />
             <label
-              htmlFor="stoneInBag"
+              htmlFor="inBag"
               className={`text-sm font-semibold dark:text-gray-200 cursor-pointer ${
-                availableStones.length === 0
+                availableFossils.length === 0
                   ? "opacity-50 cursor-not-allowed"
                   : ""
               }`}
             >
-              {t("modals.addStone.inBagLabel")}
+              {t("modals.addFossil.inBagLabel")}
             </label>
           </div>
 
           <div>
             <div className="flex items-center gap-2 mb-1">
               <label className="block text-sm font-bold text-gray-700 dark:text-gray-300">
-                {t("modals.addStone.locationLabel")}{" "}
+                {t("modals.addFossil.locationLabel")}{" "}
                 <span className="text-red-500">*</span>
               </label>
               <Tooltip
                 side="top"
-                content={t("modals.addStone.locationTooltip")}
+                content={t("modals.addFossil.locationTooltip")}
               >
                 <span
                   className="text-gray-400 hover:text-gray-600 dark:text-gray-500 dark:hover:text-gray-300 cursor-help"
-                  aria-label={t("modals.addStone.locationTooltipLabel")}
+                  aria-label={t("modals.addFossil.locationTooltipLabel")}
                 >
                   <FiInfo size={16} />
                 </span>
@@ -154,7 +163,7 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
               onChange={setLocation}
               isOpen={isOpen}
               gameVersionId={gameVersionId}
-              disabled={inBag || availableStones.length === 0}
+              disabled={inBag || availableFossils.length === 0}
               placeholder={inBag ? "" : t("common.routePlaceholder")}
             />
           </div>
@@ -169,14 +178,14 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
             </button>
             <button
               type="submit"
-              disabled={!selectedStoneId || (!inBag && !location.trim())}
+              disabled={!selectedFossilId || (!inBag && !location.trim())}
               className={`px-4 py-2 rounded-md font-semibold shadow ${focusRingClasses} ${
-                selectedStoneId && (inBag || location.trim())
+                selectedFossilId && (inBag || location.trim())
                   ? "bg-green-600 text-white hover:bg-green-700"
                   : "bg-gray-300 dark:bg-gray-600 text-gray-500 dark:text-gray-400 cursor-not-allowed"
               }`}
             >
-              {t("modals.addStone.submit")}
+              {t("modals.addFossil.submit")}
             </button>
           </div>
         </form>
@@ -185,4 +194,4 @@ const AddStoneModal: React.FC<AddStoneModalProps> = ({
   );
 };
 
-export default AddStoneModal;
+export default AddFossilModal;
