+++ src/components/modals/AddLostPokemonModal.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useId, useMemo, useRef, useState } from "react";
-import { searchPokemonNames } from "@/src/services/pokemonSearch";
-import { getSpriteUrlForPokemonName } from "@/src/services/sprites";
+import { searchPokemonNames } from "@/src/services/pokemonSearch.ts";
+import { getSpriteUrlForPokemonName } from "@/src/services/sprites.ts";
 import {
   focusRingClasses,
   focusRingInputClasses,
-} from "@/src/styles/focusRing";
-import type { Pokemon } from "@/types";
+} from "@/src/styles/focusRing.ts";
+import type { Pokemon } from "@/types.ts";
 import { useTranslation } from "react-i18next";
-import { normalizeLanguage } from "@/src/utils/language";
-import LocationSuggestionInput from "@/src/components/LocationSuggestionInput";
-import { useFocusTrap } from "@/src/hooks/useFocusTrap";
+import { normalizeLanguage } from "@/src/utils/language.ts";
+import LocationSuggestionInput from "@/src/components/LocationSuggestionInput.tsx";
+import { useFocusTrap } from "@/src/hooks/useFocusTrap.ts";
 
 interface AddLostPokemonModalProps {
   isOpen: boolean;
