+++ src/components/modals/AddStoneModal.tsx
@@ -1,57 +1,48 @@
 import React, { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { FOSSILS } from "@/src/services/init";
+import { STONES } from "@/src/services/init.ts";
 import { FiX, FiInfo } from "react-icons/fi";
-import LocationSuggestionInput from "./LocationSuggestionInput";
-import Tooltip from "./Tooltip";
-import { useFocusTrap } from "@/src/hooks/useFocusTrap";
+import LocationSuggestionInput from "../LocationSuggestionInput.tsx";
+import Tooltip from "@/src/components/basic/Tooltip.tsx";
+import { useFocusTrap } from "@/src/hooks/useFocusTrap.ts";
 import {
   focusRingCardClasses,
   focusRingClasses,
 } from "@/src/styles/focusRing.ts";
 
-interface AddFossilModalProps {
+interface AddStoneModalProps {
   isOpen: boolean;
   onClose: () => void;
-  onAdd: (fossilId: string, location: string, inBag: boolean) => void;
+  onAdd: (stoneId: string, location: string, inBag: boolean) => void;
   maxGeneration: number;
   alreadyOwnedIds: string[];
-  infiniteFossilsEnabled: boolean;
   gameVersionId?: string;
 }
 
-const AddFossilModal: React.FC<AddFossilModalProps> = ({
+const AddStoneModal: React.FC<AddStoneModalProps> = ({
   isOpen,
   onClose,
   onAdd,
   maxGeneration,
-  alreadyOwnedIds,
-  infiniteFossilsEnabled,
   gameVersionId,
 }) => {
   const { t } = useTranslation();
   const { containerRef } = useFocusTrap(isOpen);
-  const [selectedFossilId, setSelectedFossilId] = useState("");
+  const [selectedStoneId, setSelectedStoneId] = useState("");
   const [location, setLocation] = useState("");
   const [inBag, setInBag] = useState(true);
 
-  const availableFossils = useMemo(() => {
-    return FOSSILS.filter((f) => {
-      const genMatch = f.gen <= maxGeneration;
-      if (!infiniteFossilsEnabled) {
-        return genMatch && !alreadyOwnedIds.includes(f.id);
-      }
-      return genMatch;
-    });
-  }, [maxGeneration, alreadyOwnedIds, infiniteFossilsEnabled]);
+  const availableStones = useMemo(() => {
+    return STONES.filter((s) => s.gen <= maxGeneration);
+  }, [maxGeneration]);
 
   if (!isOpen) return null;
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    if (!selectedFossilId) return;
-    onAdd(selectedFossilId, inBag ? "" : location, inBag);
-    setSelectedFossilId("");
+    if (!selectedStoneId) return;
+    onAdd(selectedStoneId, inBag ? "" : location, inBag);
+    setSelectedStoneId("");
     setLocation("");
     setInBag(true);
   };
@@ -67,7 +58,7 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
       >
         <div className="flex justify-between items-center mb-4">
           <h2 className="text-xl font-bold dark:text-gray-100">
-            {t("modals.addFossil.title")}
+            {t("modals.addStone.title")}
           </h2>
           <button
             onClick={onClose}
@@ -80,37 +71,37 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
             <label className="block text-sm font-bold text-gray-700 dark:text-gray-300 mb-2">
-              {t("modals.addFossil.fossilLabel")}
+              {t("modals.addStone.stoneLabel")}
             </label>
 
-            {availableFossils.length > 0 ? (
+            {availableStones.length > 0 ? (
               <div className="grid grid-cols-3 gap-2 p-1">
-                {availableFossils.map((f) => (
+                {availableStones.map((s) => (
                   <button
-                    key={f.id}
+                    key={s.id}
                     type="button"
-                    onClick={() => setSelectedFossilId(f.id)}
+                    onClick={() => setSelectedStoneId(s.id)}
                     className={`flex flex-col items-center p-2 rounded-md border transition-all ${focusRingCardClasses} ${
-                      selectedFossilId === f.id
+                      selectedStoneId === s.id
                         ? "border-green-500 bg-green-50 dark:bg-green-900/20"
                         : "border-gray-200 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-700"
                     }`}
                   >
                     <img
-                      src={`/fossil-sprites/${f.sprite}`}
-                      alt={f.id}
+                      src={`/stone-sprites/${s.sprite}`}
+                      alt={s.id}
                       className="w-10 h-10 object-contain"
                       style={{ imageRendering: "pixelated" }}
                     />
                     <span className="text-[10px] mt-1 text-center dark:text-gray-200">
-                      {t(`fossils.${f.id}`)}
+                      {t(`stones.${s.id}`)}
                     </span>
                   </button>
                 ))}
               </div>
             ) : (
               <div className="p-4 rounded-md border border-amber-200 dark:border-amber-800 bg-amber-50 dark:bg-amber-900/20 text-xs text-amber-800 dark:text-amber-200 leading-relaxed">
-                {t("modals.addFossil.emptyListExplanation")}
+                {t("modals.addStone.emptyListExplanation")}
               </div>
             )}
           </div>
@@ -118,40 +109,40 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
           <div className="flex items-center gap-2 py-2">
             <input
               type="checkbox"
-              id="inBag"
+              id="stoneInBag"
               checked={inBag}
               onChange={(e) => {
                 setInBag(e.target.checked);
                 if (e.target.checked) setLocation("");
               }}
-              disabled={availableFossils.length === 0}
+              disabled={availableStones.length === 0}
               className="h-4 w-4 accent-green-600 cursor-pointer disabled:cursor-not-allowed"
             />
             <label
-              htmlFor="inBag"
+              htmlFor="stoneInBag"
               className={`text-sm font-semibold dark:text-gray-200 cursor-pointer ${
-                availableFossils.length === 0
+                availableStones.length === 0
                   ? "opacity-50 cursor-not-allowed"
                   : ""
               }`}
             >
-              {t("modals.addFossil.inBagLabel")}
+              {t("modals.addStone.inBagLabel")}
             </label>
           </div>
 
           <div>
             <div className="flex items-center gap-2 mb-1">
               <label className="block text-sm font-bold text-gray-700 dark:text-gray-300">
-                {t("modals.addFossil.locationLabel")}{" "}
+                {t("modals.addStone.locationLabel")}{" "}
                 <span className="text-red-500">*</span>
               </label>
               <Tooltip
                 side="top"
-                content={t("modals.addFossil.locationTooltip")}
+                content={t("modals.addStone.locationTooltip")}
               >
                 <span
                   className="text-gray-400 hover:text-gray-600 dark:text-gray-500 dark:hover:text-gray-300 cursor-help"
-                  aria-label={t("modals.addFossil.locationTooltipLabel")}
+                  aria-label={t("modals.addStone.locationTooltipLabel")}
                 >
                   <FiInfo size={16} />
                 </span>
@@ -163,7 +154,7 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
               onChange={setLocation}
               isOpen={isOpen}
               gameVersionId={gameVersionId}
-              disabled={inBag || availableFossils.length === 0}
+              disabled={inBag || availableStones.length === 0}
               placeholder={inBag ? "" : t("common.routePlaceholder")}
             />
           </div>
@@ -178,14 +169,14 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
             </button>
             <button
               type="submit"
-              disabled={!selectedFossilId || (!inBag && !location.trim())}
+              disabled={!selectedStoneId || (!inBag && !location.trim())}
               className={`px-4 py-2 rounded-md font-semibold shadow ${focusRingClasses} ${
-                selectedFossilId && (inBag || location.trim())
+                selectedStoneId && (inBag || location.trim())
                   ? "bg-green-600 text-white hover:bg-green-700"
                   : "bg-gray-300 dark:bg-gray-600 text-gray-500 dark:text-gray-400 cursor-not-allowed"
               }`}
             >
-              {t("modals.addFossil.submit")}
+              {t("modals.addStone.submit")}
             </button>
           </div>
         </form>
@@ -194,4 +185,4 @@ const AddFossilModal: React.FC<AddFossilModalProps> = ({
   );
 };
 
-export default AddFossilModal;
+export default AddStoneModal;
