+++ src/components/modals/DeleteTrackerModal.tsx
@@ -1,8 +1,8 @@
 import React, { useId } from "react";
 import { FiAlertTriangle, FiX } from "react-icons/fi";
 import { useTranslation } from "react-i18next";
-import { useFocusTrap } from "@/src/hooks/useFocusTrap";
-import { focusRingClasses } from "@/src/styles/focusRing";
+import { useFocusTrap } from "@/src/hooks/useFocusTrap.ts";
+import { focusRingClasses } from "@/src/styles/focusRing.ts";
 
 interface DeleteTrackerModalProps {
   isOpen: boolean;
