+++ src/components/pages/LoginPage.tsx
@@ -1,63 +1,33 @@
 import React, { useState } from "react";
-import { createUserWithEmailAndPassword } from "firebase/auth";
-import type { FirebaseError } from "firebase/app";
-import { auth } from "../firebaseConfig";
+import { signInWithEmailAndPassword } from "firebase/auth";
+import { auth } from "../../firebaseConfig.ts";
 import {
   focusRingBlueClasses,
   focusRingClasses,
   focusRingInputClasses,
-} from "@/src/styles/focusRing";
+} from "@/src/styles/focusRing.ts";
 import { useTranslation } from "react-i18next";
 
-type RegisterPageProps = {
-  onSwitchToLogin: () => void;
+type LoginPageProps = {
+  onSwitchToRegister: () => void;
 };
 
-const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
+const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
   const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
   const { t } = useTranslation();
 
-  const resolveErrorMessage = (err: unknown): string => {
-    if (typeof err === "object" && err !== null && "code" in err) {
-      const code = (err as FirebaseError).code;
-      switch (code) {
-        case "auth/email-already-in-use":
-          return t("auth.register.errors.emailInUse");
-        case "auth/invalid-email":
-          return t("auth.register.errors.invalidEmail");
-        case "auth/weak-password":
-          return t("auth.register.errors.weakPassword");
-        default:
-          break;
-      }
-    }
-    return t("auth.register.errors.general");
-  };
-
-  const handleRegister = async (e: React.FormEvent) => {
+  const handleAuthAction = async (e: React.FormEvent) => {
     e.preventDefault();
     setError(null);
-
-    if (password.length < 8) {
-      setError(t("auth.register.errors.weakPassword"));
-      return;
-    }
-
-    if (password !== confirmPassword) {
-      setError(t("auth.register.passwordMismatch"));
-      return;
-    }
-
     setLoading(true);
     try {
-      await createUserWithEmailAndPassword(auth, email.trim(), password);
+      await signInWithEmailAndPassword(auth, email, password);
     } catch (err) {
-      console.error("Error registering user", err);
-      setError(resolveErrorMessage(err));
+      console.error("Error signing in", err);
+      setError(t("auth.login.error"));
     } finally {
       setLoading(false);
     }
@@ -74,27 +44,27 @@ const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
               className="mx-auto mb-3 w-40 h-40 object-contain"
             />
             <h1 className="text-2xl sm:text-3xl font-bold font-press-start tracking-tighter dark:text-gray-100">
-              {t("auth.register.title")}
+              {t("auth.login.title")}
             </h1>
             <p className="text-sm text-gray-500 dark:text-gray-400 mt-2 text-center">
-              <span className="block">{t("auth.register.headline")}</span>
+              <span className="block">{t("auth.login.headline")}</span>
               <span className="block mt-1">
-                {t("auth.register.haveAccount")}{" "}
+                {t("auth.login.fallback")}{" "}
                 <button
                   type="button"
-                  onClick={onSwitchToLogin}
+                  onClick={onSwitchToRegister}
                   className={`font-semibold text-blue-600 hover:text-blue-700 dark:text-blue-400 hover:underline focus-visible:underline ${focusRingBlueClasses}`}
                 >
-                  {t("auth.register.toLogin")}
+                  {t("auth.login.registerPrompt")}
                 </button>
               </span>
             </p>
           </header>
 
-          <form onSubmit={handleRegister} className="mt-6 space-y-4">
+          <form onSubmit={handleAuthAction} className="mt-6 space-y-4">
             <div>
               <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("auth.register.emailLabel")}
+                {t("auth.login.emailLabel")}
               </label>
               <input
                 type="email"
@@ -108,30 +78,16 @@ const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
             </div>
             <div>
               <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("auth.register.passwordLabel")}
+                {t("auth.login.passwordLabel")}
               </label>
               <input
                 type="password"
-                autoComplete="new-password"
+                autoComplete="current-password"
                 required
                 className={`w-full border border-gray-300 dark:border-gray-600 rounded-lg px-3 py-2 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 ${focusRingInputClasses}`}
                 placeholder="••••••••"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
-              />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("auth.register.confirmPasswordLabel")}
-              </label>
-              <input
-                type="password"
-                autoComplete="new-password"
-                required
-                className={`w-full border border-gray-300 dark:border-gray-600 rounded-lg px-3 py-2 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 ${focusRingInputClasses}`}
-                placeholder="••••••••"
-                value={confirmPassword}
-                onChange={(e) => setConfirmPassword(e.target.value)}
               />
             </div>
 
@@ -146,9 +102,7 @@ const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
               disabled={loading}
               className={`w-full bg-green-600 text-white font-semibold py-2 px-6 rounded-lg hover:bg-green-700 transition-colors shadow-md disabled:opacity-70 ${focusRingClasses}`}
             >
-              {loading
-                ? `${t("auth.register.submit")}…`
-                : t("auth.register.submit")}
+              {loading ? `${t("auth.login.submit")}…` : t("auth.login.submit")}
             </button>
           </form>
         </div>
@@ -157,4 +111,4 @@ const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
   );
 };
 
-export default RegisterPage;
+export default LoginPage;
