+++ src/components/pages/RegisterPage.tsx
@@ -1,33 +1,63 @@
 import React, { useState } from "react";
-import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../firebaseConfig";
+import { createUserWithEmailAndPassword } from "firebase/auth";
+import type { FirebaseError } from "firebase/app";
+import { auth } from "../../firebaseConfig.ts";
 import {
   focusRingBlueClasses,
   focusRingClasses,
   focusRingInputClasses,
-} from "@/src/styles/focusRing";
+} from "@/src/styles/focusRing.ts";
 import { useTranslation } from "react-i18next";
 
-type LoginPageProps = {
-  onSwitchToRegister: () => void;
+type RegisterPageProps = {
+  onSwitchToLogin: () => void;
 };
 
-const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
+const RegisterPage: React.FC<RegisterPageProps> = ({ onSwitchToLogin }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
   const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
   const { t } = useTranslation();
 
-  const handleAuthAction = async (e: React.FormEvent) => {
+  const resolveErrorMessage = (err: unknown): string => {
+    if (typeof err === "object" && err !== null && "code" in err) {
+      const code = (err as FirebaseError).code;
+      switch (code) {
+        case "auth/email-already-in-use":
+          return t("auth.register.errors.emailInUse");
+        case "auth/invalid-email":
+          return t("auth.register.errors.invalidEmail");
+        case "auth/weak-password":
+          return t("auth.register.errors.weakPassword");
+        default:
+          break;
+      }
+    }
+    return t("auth.register.errors.general");
+  };
+
+  const handleRegister = async (e: React.FormEvent) => {
     e.preventDefault();
     setError(null);
+
+    if (password.length < 8) {
+      setError(t("auth.register.errors.weakPassword"));
+      return;
+    }
+
+    if (password !== confirmPassword) {
+      setError(t("auth.register.passwordMismatch"));
+      return;
+    }
+
     setLoading(true);
     try {
-      await signInWithEmailAndPassword(auth, email, password);
+      await createUserWithEmailAndPassword(auth, email.trim(), password);
     } catch (err) {
-      console.error("Error signing in", err);
-      setError(t("auth.login.error"));
+      console.error("Error registering user", err);
+      setError(resolveErrorMessage(err));
     } finally {
       setLoading(false);
     }
@@ -44,27 +74,27 @@ const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
               className="mx-auto mb-3 w-40 h-40 object-contain"
             />
             <h1 className="text-2xl sm:text-3xl font-bold font-press-start tracking-tighter dark:text-gray-100">
-              {t("auth.login.title")}
+              {t("auth.register.title")}
             </h1>
             <p className="text-sm text-gray-500 dark:text-gray-400 mt-2 text-center">
-              <span className="block">{t("auth.login.headline")}</span>
+              <span className="block">{t("auth.register.headline")}</span>
               <span className="block mt-1">
-                {t("auth.login.fallback")}{" "}
+                {t("auth.register.haveAccount")}{" "}
                 <button
                   type="button"
-                  onClick={onSwitchToRegister}
+                  onClick={onSwitchToLogin}
                   className={`font-semibold text-blue-600 hover:text-blue-700 dark:text-blue-400 hover:underline focus-visible:underline ${focusRingBlueClasses}`}
                 >
-                  {t("auth.login.registerPrompt")}
+                  {t("auth.register.toLogin")}
                 </button>
               </span>
             </p>
           </header>
 
-          <form onSubmit={handleAuthAction} className="mt-6 space-y-4">
+          <form onSubmit={handleRegister} className="mt-6 space-y-4">
             <div>
               <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("auth.login.emailLabel")}
+                {t("auth.register.emailLabel")}
               </label>
               <input
                 type="email"
@@ -78,16 +108,30 @@ const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
             </div>
             <div>
               <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("auth.login.passwordLabel")}
+                {t("auth.register.passwordLabel")}
               </label>
               <input
                 type="password"
-                autoComplete="current-password"
+                autoComplete="new-password"
                 required
                 className={`w-full border border-gray-300 dark:border-gray-600 rounded-lg px-3 py-2 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 ${focusRingInputClasses}`}
                 placeholder="••••••••"
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
+              />
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
+                {t("auth.register.confirmPasswordLabel")}
+              </label>
+              <input
+                type="password"
+                autoComplete="new-password"
+                required
+                className={`w-full border border-gray-300 dark:border-gray-600 rounded-lg px-3 py-2 bg-white dark:bg-gray-700 text-gray-900 dark:text-gray-100 ${focusRingInputClasses}`}
+                placeholder="••••••••"
+                value={confirmPassword}
+                onChange={(e) => setConfirmPassword(e.target.value)}
               />
             </div>
 
@@ -102,7 +146,9 @@ const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
               disabled={loading}
               className={`w-full bg-green-600 text-white font-semibold py-2 px-6 rounded-lg hover:bg-green-700 transition-colors shadow-md disabled:opacity-70 ${focusRingClasses}`}
             >
-              {loading ? `${t("auth.login.submit")}…` : t("auth.login.submit")}
+              {loading
+                ? `${t("auth.register.submit")}…`
+                : t("auth.register.submit")}
             </button>
           </form>
         </div>
@@ -111,4 +157,4 @@ const LoginPage: React.FC<LoginPageProps> = ({ onSwitchToRegister }) => {
   );
 };
 
-export default LoginPage;
+export default RegisterPage;
